backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
pytest==7.4.3
pytest-flask==1.3.0
